Update README.md
This is my project on T cell lineage tracking in microscopic movies. The algorithm includes 3 steps: 1. Tracking 2. Segmentation 3. Division detection.
Description of the files presented here:

TRACK-4.py. This is a 3D CNN for tracking T cells in fluorescent images. It is suitable for sections of cell movies containing 4 cells only. There are also similar neural networks specializing in tracking 1,2,3 and 5 cells but they are not presented here as they are very similar.
SEGMENTOR.py. This is U-Net for segmenting 96 x 96 patches cropped out of the frames. This U-Net learns to segment a cell of interest-only rather than the whole patch.
DIVISION DETECTOR.py
EXECUTION.py
An example training sample for TRACKER-4.
An example training sample for SEGMENTOR.
An example of applying the algorithm to a T cell movie.